requirements.txt
# pip install -r requirements.txt
# utilities
numpy
pandas

#face recognition
scikit-learn
dlib-bin

git+https://github.com/ageitgey/face_recognition_models
setuptools<70.0.0

supabase   #database
bcrypt  #password

segno   #QR code
pillow  #image preprocessing 

streamlit #frontend
librosa
resemblyzer
